Hardcode the /etc/live/disable-rc.d/ directory
This allows us to check if it exists without loading the lib so we can
bail early. Also use the existence of that lib as the sole criterion for
running.  This is faster and allows us to recover if the boot is
aborted and services have not yet been restored.