Improve handling of blkid output.
We should always be able to get the fs type regardless of wacky
characters in the label.  Very malicous labels could spoof the
uuid.  Oh well.  Some malicious labels will be ignored.

We don't handle labels with spaces correctly in any_to_dev() but
you can't specify spaces in boot parameters so this may be moot.