Update live-init to usually make an xorg.conf with vesa driver
You need to /sbin/make-xorg-conf program for this to work properly.
It is in the Xtra-files repo.  The problem is that if the framebuffer
is enable and there is no hardwware specific video driver then X
now defaults to using fbdev at the resolutions of the consoles.
We want the consoles to default to 800x600 so this is a problem.

Therefore in cases where this seems it will be a problem, we
generate an xorg.conf that specifices the vesa driver.  This
is working in vbox but I have not tested it on a laptop yet.