Change xorg boot parameters and logic for creating xorg.conf
Use the name of the fraoebuffer instead of loaded modules for
detecting graphics drivers.  The framebuffer is the key.  Make
a safe xorg.conf in "failsafe" mode. Erase old xorg.conf if there
is a modeset driver in play.  This is far from perfect but it
might do the greatest good for the greatest number.

It would be safest to always erease the previous xorg.conf but
I don't want to deny people the ability to have a custom one.
In addition "noxorg" will now prevent us from erasing xorg.conf
in addtion to preventing us writing a new one.

Should backup xorg.conf before erasing it.