Handle default user and X vs. not-X better
Don't assume default user is "demo" Get the default from
slim.conf and lightdm.conf first. Then UID=1000 and then
finally "demo" if all of the above fails.

Group the X code better so all the user specific stuff is in
one place and we can return before running it if we can't find
a default user.