Change how we deal with exiting framebuffer and virtualbox
Disable virtualbox mode if $no_default gets set for any reason.
Disable safe_mode and set no_default if *any* framebuffer is
found except the vga framebuffer.  This logic may be wrong
but the idea is that if a framebuffer exists this early then
it is probably set to the max resolution (like during an UEFI
boot) so allowing the fbdev default is not a bad thing.  If
the framebuffer is controled by a KMS driver then any safe
or default vesa xorg.conf will keep X from starting.