Replace hotplug with a bg process that does cold plug in a loop
Big *sigh*.   I don't know why they keep making this so hard but
I have my usual suspects.  I start and stop the coldplug loop
twice.  Once inside of main_core() and again outside of it.  Starting
it inside and stopping it outside caused a big hang.  I'm sure I
could fix that but this was faster and easier.  Most likely it will
rarely be needed once we've mounted the boot device and persistence
devide but ISTM it is best to run it until the very end even if it
is only marginally better.