Let dev names be up to 10 chars long.  Deal with size and model better.
This gives us one extra char for mmcblk0p1 names which is 9 chars
long.  I hope this is enough.  I will have wrapping problems if
I have to deal with names that are much longer than this.

We continue if there is no size.  We truncate the model name to
16 chars.  I haven't seen one longer than this but truncating will
help keep the columns aligned if there is one longer.